requirements.txt
appnope==0.1.0
asgiref==3.2.10
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
Django==3.0.5
django-redis==4.12.1
django-rest-framework==0.1.0
djangorestframework==3.12.1
djongo==1.3.3
idna==2.10
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.1
pymongo==3.11.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
sqlparse==0.2.4
traitlets==5.0.5
urllib3==1.25.10
wcwidth==0.2.5
